+++ lib/test_prof/recipes/rspec/factory_all_stub.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+require "test_prof/factory_all_stub"
+
+TestProf::FactoryAllStub.init
+
+RSpec.shared_context "factory:stub", factory: :stub do
+  prepend_before(:all) { TestProf::FactoryAllStub.enable! }
+  append_after(:all) { TestProf::FactoryAllStub.disable! }
+end