+++ spec/integrations/factory_all_stub_spec.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+require "spec_helper"
+
+describe "FactoryAllStub" do
+  specify "it works" do
+    output = run_rspec('factory_all_stub')
+
+    expect(output).to include("3 examples, 0 failures")
+  end
+end