+++ spec/support/transactional_context.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require "active_record"
+
 if ActiveRecord::VERSION::MAJOR < 4
   require "test_prof/ext/active_record_3"
   using TestProf::ActiveRecord3Transactions